删除存在子节点的接口时进行判断是否存在子节点

+++ server/crates/service_hub/src/perm/service/dept.rs
@@ -139,6 +139,22 @@ impl<'a> DeptService<'a> {
 
     /// 删除数据
     pub async fn delete(&self, id: i32) -> Result<u64, ErrorMsg> {
+        let dept_children = self.dept_dao.children(id).await.map_err(|err| {
+            error!("获取所有子列表失败, err: {:#?}", err);
+            Error::DbQueryError
+                .into_msg()
+                .with_msg("获取所有子列表失败")
+        })?;
+        if !dept_children.is_empty() {
+            error!(
+                "请先删除子列表, children count: {:#?}",
+                dept_children.len()
+            );
+            return Err(Error::DbDataExistChildrenError
+                .into_msg()
+                .with_msg("请先删除子列表"));
+        }
+
         let result = self.dept_dao.delete(id).await.map_err(|err| {
             error!("删除部门信息失败, err: {:#?}", err);
             Error::DbDeleteError.into_msg().with_msg("删除部门信息失败")
